fix(financials): address PR #592 round 8 — extract utils, dedupe formatter & SEO call
- move isEmptySnapshot/fetchAndApplyQuarterSnapshot/FetchQuarterContext from hooks/useFinancialsPeriod.ts to utils/financialsPeriodUtils.ts (non-hook utils belong in utils/, §18)
- extract shared usdFormatter to utils/numberFormat.ts; AxisScoreCard + StatementTable import it (no duplicate definition, §1)
- page.tsx breadcrumb: reuse already-extracted url instead of re-calling buildSymbolFinancialsSeoContent (§2)
- skip S1 (import type React): false positive — React.CSSProperties is used
